refactor(native): move runtime engine core into subdomain folder
