+++ src/native/engine/README.md
@@ -101,8 +101,8 @@ Desktop shell (Steam Windows client)
 Runtime responsibilities are split by domain service modules rather than one
 monolithic `engine.c` control block:
 
-- `runtime/engine_lifecycle.*`: bootstrap/teardown orchestration.
-- `runtime/engine_tick.*`: tick pipeline execution and phase ordering.
+- `runtime/engine/engine_lifecycle.*`: bootstrap/teardown orchestration.
+- `runtime/engine/engine_tick.*`: tick pipeline execution and phase ordering.
 - `runtime/input/input_contract.*`: click/move intent contract and sanitization.
 - `runtime/move_target_domain.*`: strongly-typed click-target state and steering decisions.
 - `runtime/render_submit.*` + `RendererDrawCommand`: typed scene command
