+++ src/native/engine/engine.c
@@ -4,14 +4,14 @@
 #include "ai/npc_merchant.h"
 #include "ai/wildlife_controller.h"
 #include "runtime/controller/controller_sync.h"
-#include "runtime/application_service_ports.h"
-#include "runtime/engine_aux_abi.h"
-#include "runtime/engine_composition.h"
-#include "runtime/engine_host.h"
-#include "runtime/engine_state.h"
-#include "runtime/engine_lifecycle.h"
-#include "runtime/engine_tick.h"
-#include "runtime/gameplay_service.h"
+#include "runtime/engine/application_service_ports.h"
+#include "runtime/engine/engine_aux_abi.h"
+#include "runtime/engine/engine_composition.h"
+#include "runtime/engine/engine_host.h"
+#include "runtime/engine/engine_state.h"
+#include "runtime/engine/engine_lifecycle.h"
+#include "runtime/engine/engine_tick.h"
+#include "runtime/engine/gameplay_service.h"
 #include "runtime/camera_follow.h"
 #include "runtime/camera_basis.h"
 #include "runtime/move_target_service.h"
