+++ src/native/engine/runtime/engine/application_service_ports.c
@@ -1,12 +1,12 @@
 #include "application_service_ports.h"
 
-#include "move_target_service.h"
-#include "player/player_runtime_service.h"
-#include "render_material.h"
-#include "render_submit.h"
-#include "terrain/terrain_height.h"
-#include "terrain/terrain_host.h"
-#include "terrain/terrain_mutation.h"
+#include "../move_target_service.h"
+#include "../player/player_runtime_service.h"
+#include "../render_material.h"
+#include "../render_submit.h"
+#include "../terrain/terrain_height.h"
+#include "../terrain/terrain_host.h"
+#include "../terrain/terrain_mutation.h"
 
 #include <string.h>
