+++ src/native/engine/runtime/engine/application_service_ports.h
@@ -2,8 +2,8 @@
 #define BANANA_ENGINE_RUNTIME_APPLICATION_SERVICE_PORTS_H
 
 #include "engine_state.h"
-#include "merchant/merchant_service.h"
-#include "player/player_motion.h"
+#include "../merchant/merchant_service.h"
+#include "../player/player_motion.h"
 
 #ifdef __cplusplus
 extern "C"
