+++ src/native/engine/runtime/engine/engine_lifecycle.h
@@ -1,10 +1,10 @@
 #ifndef BANANA_ENGINE_RUNTIME_ENGINE_LIFECYCLE_H
 #define BANANA_ENGINE_RUNTIME_ENGINE_LIFECYCLE_H
 
-#include "../ai/controller.h"
-#include "../world/world.h"
-#include "player/player_registry.h"
-#include "terrain/terrain_runtime.h"
+#include "../../ai/controller.h"
+#include "../../world/world.h"
+#include "../player/player_registry.h"
+#include "../terrain/terrain_runtime.h"
 
 #ifdef __cplusplus
 extern "C"
