+++ src/native/engine/runtime/engine/engine_state.h
@@ -1,14 +1,14 @@
 #ifndef BANANA_ENGINE_RUNTIME_ENGINE_STATE_H
 #define BANANA_ENGINE_RUNTIME_ENGINE_STATE_H
 
-#include "../ai/controller.h"
-#include "../physics/world.h"
-#include "../render/mesh.h"
-#include "../render/renderer.h"
-#include "../render/window.h"
-#include "../world/world.h"
-#include "move_target_domain.h"
-#include "terrain/terrain_runtime.h"
+#include "../../ai/controller.h"
+#include "../../physics/world.h"
+#include "../../render/mesh.h"
+#include "../../render/renderer.h"
+#include "../../render/window.h"
+#include "../../world/world.h"
+#include "../move_target_domain.h"
+#include "../terrain/terrain_runtime.h"
 
 #ifdef __cplusplus
 extern "C"
