+++ src/native/engine/runtime/engine/gameplay_service.h
@@ -1,8 +1,8 @@
 #ifndef BANANA_ENGINE_RUNTIME_GAMEPLAY_SERVICE_H
 #define BANANA_ENGINE_RUNTIME_GAMEPLAY_SERVICE_H
 
-#include "../ai/controller.h"
-#include "../world/world.h"
+#include "../../ai/controller.h"
+#include "../../world/world.h"
 
 #ifdef __cplusplus
 extern "C"
