+++ tests/native/architecture/runtime_architecture_guard_rules.h
@@ -41,7 +41,7 @@ static const FileRule kRuntimeArchitectureGuardRules[] = {
         1,
     },
     {
-        "runtime/gameplay_service.c",
+        "runtime/engine/gameplay_service.c",
         "domain services must not depend on orchestration",
         {"orchestration/"},
         1,
@@ -65,38 +65,38 @@ static const FileRule kRuntimeArchitectureGuardRules[] = {
         1,
     },
     {
-        "runtime/engine_composition.c",
+        "runtime/engine/engine_composition.c",
         "composition should keep service ports in explicit state, not hidden globals",
         {"static const RuntimeApplicationServicePorts *s_service_ports"},
         1,
     },
     {
-        "runtime/engine_tick.c",
+        "runtime/engine/engine_tick.c",
         "tick orchestration should delegate click normalization policy",
         {"(click_x / (float)input_width) * 2.f - 1.f",
          "1.f - (click_y / (float)input_height) * 2.f"},
         2,
     },
     {
-        "runtime/engine_tick.c",
+        "runtime/engine/engine_tick.c",
         "tick orchestration should delegate terrain budget policy",
         {"runtime_phase_terrain_budget(context, terrain_rebuild, 2)"},
         1,
     },
     {
-        "runtime/engine_tick.c",
+        "runtime/engine/engine_tick.c",
         "tick orchestration should delegate right-click polling to tick input phase",
         {"window_take_right_click("},
         1,
     },
     {
-        "runtime/engine_tick.c",
+        "runtime/engine/engine_tick.c",
         "tick orchestration should delegate post-gameplay sequencing to tick post phase",
         {"runtime_sync_controller_positions("},
         1,
     },
     {
-        "runtime/engine_tick.c",
+        "runtime/engine/engine_tick.c",
         "engine tick adapter should delegate sequencing to runtime tick orchestration",
         {"runtime_tick_input_phase_process(",
          "runtime_phase_viewport_resize(",
