+++ tests/native/runtime/tick/runtime_engine_tick_test.c
@@ -1,4 +1,4 @@
-#include "runtime/engine_tick.h"
+#include "runtime/engine/engine_tick.h"
 
 #include <stdio.h>
