+++ tests/native/runtime_application_service_ports_test.c
@@ -1,5 +1,5 @@
-#include "runtime/application_service_ports.h"
-#include "runtime/engine_state.h"
+#include "runtime/engine/application_service_ports.h"
+#include "runtime/engine/engine_state.h"
 
 #include <math.h>
 #include <stdio.h>
