+++ tests/native/runtime_engine_aux_abi_test.c
@@ -1,4 +1,4 @@
-#include "runtime/engine_aux_abi.h"
+#include "runtime/engine/engine_aux_abi.h"
 
 #include <stdio.h>
 #include <stdlib.h>
