+++ tests/native/runtime_engine_host_test.c
@@ -1,4 +1,4 @@
-#include "runtime/engine_host.h"
+#include "runtime/engine/engine_host.h"
 
 #include <stdio.h>
