+++ tests/native/runtime_engine_lifecycle_test.c
@@ -1,4 +1,4 @@
-#include "runtime/engine_lifecycle.h"
+#include "runtime/engine/engine_lifecycle.h"
 
 #include <math.h>
 #include <stdio.h>
