+++ tests/native/runtime_engine_state_test.c
@@ -1,4 +1,4 @@
-#include "runtime/engine_state.h"
+#include "runtime/engine/engine_state.h"
 
 #include <stdio.h>
