+++ tests/native/runtime_gameplay_service_test.c
@@ -1,4 +1,4 @@
-#include "runtime/gameplay_service.h"
+#include "runtime/engine/gameplay_service.h"
 
 #include <stdio.h>
 #include <string.h>
