+++ tests/native/runtime_input_abi_test.c
@@ -1,4 +1,4 @@
-#include "runtime/engine_state.h"
+#include "runtime/engine/engine_state.h"
 #include "runtime/input/input_abi.h"
 #include "runtime/input/input_contract.h"
